+++ CHANGELOG.rst
@@ -1,6 +1,11 @@
 Change Log
 ==========
 
+UNRELEASED
+----------
+* :star: Added the option not to include PDF's inline inside the PDF. The PDF will return a zip with the PDF and all attachments that cannot be put inline inside an attachment folder. (#1448)
+* :+1: Also accept a 202 Accepted HTTP response for the scope clone.(#1459)
+
 v4.12.0 (2JUL24)
 ----------------
 * :+1: We've refactored the way we handle the retrieval of descendants of a `Part`, by properly targeting the root to be removed from the list. (#1444)