+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ Change Log
 UNRELEASED
 ----------
 * :star: Added the option not to include PDF's inline inside the PDF. The PDF will return a zip with the PDF and all attachments that cannot be put inline inside an attachment folder. (#1448)
+* :+1: Also accept a 202 Accepted HTTP response for the scope clone and improved API error handling.(#1459)
 
 v4.12.0 (2JUL24)
 ----------------